+++ nucliadb_ingest/nucliadb_ingest/tests/fixtures.py
@@ -52,13 +52,19 @@ from nucliadb_protos import resources_pb2 as rpb
 from nucliadb_protos import utils_pb2 as upb
 from nucliadb_protos import writer_pb2_grpc
 from nucliadb_utils.audit.basic import BasicAuditStorage
+from nucliadb_utils.audit.stream import StreamAuditStorage
 from nucliadb_utils.cache.redis import RedisPubsub
 from nucliadb_utils.cache.settings import settings as cache_settings
 from nucliadb_utils.cache.utility import Cache
 from nucliadb_utils.indexing import IndexingUtility
 from nucliadb_utils.settings import indexing_settings, transaction_settings
 from nucliadb_utils.storages.settings import settings as storage_settings
-from nucliadb_utils.utilities import Utility, clear_global_cache, set_utility
+from nucliadb_utils.utilities import (
+    Utility,
+    clear_global_cache,
+    get_utility,
+    set_utility,
+)
 
 images.settings["nucliadb_node_reader"] = {
     "image": "eu.gcr.io/stashify-218417/node",
@@ -182,6 +188,14 @@ async def processor(redis_driver, gcs_storage, cache, audit):
 
 
 @pytest.fixture(scope="function")
+async def stream_processor(redis_driver, gcs_storage, cache, stream_audit):
+    proc = Processor(redis_driver, gcs_storage, stream_audit, cache, 1)
+    await proc.initialize()
+    yield proc
+    await proc.finalize()
+
+
+@pytest.fixture(scope="function")
 async def local_files():
     storage_settings.local_testing_files = f"{dirname(__file__)}"
 
@@ -260,6 +274,7 @@ async def redis_driver(redis):
     driver = RedisDriver(url=url)
     await driver.initialize()
     await driver.redis.flushall()
+    print(f"Redis driver ready at {url}")
     yield driver
     await driver.finalize()
     settings.driver_redis_url = None
@@ -274,14 +289,20 @@ async def txn(redis_driver):
 
 @pytest.fixture(scope="function")
 async def cache(redis):
-    url = f"redis://{redis[0]}:{redis[1]}"
-    pubsub = RedisPubsub(url)
-    await pubsub.initialize()
+
+    pubsub = get_utility(Utility.PUBSUB)
+    if pubsub is None:
+        url = f"redis://{redis[0]}:{redis[1]}"
+        pubsub = RedisPubsub(url)
+        await pubsub.initialize()
+        set_utility(Utility.PUBSUB, pubsub)
+
     che = Cache(pubsub)
     await che.initialize()
     yield che
     await che.finalize()
     await pubsub.finalize()
+    set_utility(Utility.PUBSUB, None)
 
 
 @pytest.fixture(scope="function")
@@ -577,6 +598,21 @@ async def knowledgebox(redis_driver: RedisDriver):
 @pytest.fixture(scope="function")
 async def audit():
     return BasicAuditStorage()
+
+
+@pytest.fixture(scope="function")
+async def stream_audit(natsd: str):
+    from nucliadb_utils.settings import audit_settings
+
+    audit = StreamAuditStorage(
+        [natsd],
+        audit_settings.audit_jetstream_target,  # type: ignore
+        audit_settings.audit_partitions,
+        audit_settings.audit_hash_seed,
+    )
+    await audit.initialize()
+    yield audit
+    await audit.finalize()
 
 
 @pytest.fixture(scope="function")