+++ nucliadb_protos/python/CHANGELOG.rst
@@ -4,7 +4,7 @@ CHANGELOG
 0.0.28 (unreleased)
 -------------------
 
-- Nothing changed yet.
+- Extend audit to provide insight of how many fields have changed and it's file fields storage movements
 
 
 0.0.27 (2022-08-03)